+++ src/routes/config/+page.svelte
@@ -20,6 +20,7 @@
 		gotissuesMode: 'Both',
 		gotissuesIssuesCount: '5',
 		gotissuesPackCount: 'All',
+		gotissuesMinPack: '0',
 		rocTop: '100',
 		rocDays: '30',
 		rocSpecific: '',
@@ -189,6 +190,11 @@
 			label="Pack Count"
 			bind:bindValue={localStorageObject.gotissuesPackCount}
 			items={['All', '1', '2', '3', '4', '5', '6', '7', '8', '9']} />
+		<FormSelect
+			id="minPack"
+			label="Minimum Pack Count"
+			bind:bindValue={localStorageObject.gotissuesMinPack}
+			items={['0', '1', '2', '3', '4', '5', '6', '7', '8', '9']} />
 		<h2 class="text-center text-2xl font-bold tracking-tight">Rate of Change</h2>
 		<FormInput label="Top {localStorageObject.rocTop}" bind:bindValue={localStorageObject.rocTop} id="top" />
 		<FormInput label="Over {localStorageObject.rocDays} Days" bind:bindValue={localStorageObject.rocDays} id="days" />